# 8296602: RISC-V: improve performance of copy_memory stub